Working in animationSystem

+++ ext_libs/include/GT/Config.hpp
@@ -3,6 +3,6 @@
 #define  CONFIG_INC
 
 typedef float Real;
-typedef long Integer;
+typedef int Integer;
 
 #endif   // ----- #ifndef CONFIG_INC  -----
